Inject planet API in handlers, start unit test refactor
In order to be tested without hitting the live Planet API,
the handlers needed to get the base API URL from elsewhere.
This was accomplished by using the http.Handler rather than
http.HandlerFunc interface, which allows a configurable Handler
object. This Handler object stores the Planet API URL, which
can be set appropriately depending on whether we are actually
serving, or testing.

This commit also starts an overhaul of `handlers_test.go`,
which had problems in that it was one huge test (as opposed to
more idiomatic small tests), its code was hitting the live
Planet API, and it unnecessarily duplicated the net/http/httptest
package in some ways.